+++ src/commands/dev/blacklist.js
@@ -7,7 +7,7 @@ const {
   ButtonStyle,
 } = require('discord.js');
 const { writeAt, toFixed, generateID } = require('../../functions/helper.js');
-const { errorMessage } = require('../../logger.js');
+const { errorMessage } = require('../../functions/logger.js');
 const config = require('../../../config.json');
 const fs = require('fs');
