+++ src/commands/general/guild.js
@@ -3,7 +3,7 @@ const { generateGuild } = require('../../functions/generateImage.js');
 const { registerGuild } = require('../../api/pixelicAPI.js');
 const { generateID } = require('../../functions/helper.js');
 const { getGuild } = require('../../api/wynnCraftAPI.js');
-const { errorMessage } = require('../../logger.js');
+const { errorMessage } = require('../../functions/logger.js');
 const config = require('../../../config.json');
 
 module.exports = {
