+++ src/commands/general/reportBug.js
@@ -10,7 +10,7 @@ const {
   Events,
 } = require('discord.js');
 const { writeAt, toFixed, generateID } = require('../../functions/helper.js');
-const { errorMessage } = require('../../logger.js');
+const { errorMessage } = require('../../functions/logger.js');
 const config = require('../../../config.json');
 
 module.exports = {
